[#85] Added Abstract / Swappable models
