+++ plans/apps.py
@@ -6,3 +6,7 @@ from . import conf as app_settings
 class PlansConfig(AppConfig):
     name = 'plans'
     verbose_name = app_settings.APP_VERBOSE_NAME
+
+    def ready(self):
+        # noinspection PyUnresolvedReferences
+        import plans.listeners  # noqa
