+++ plans/views.py
@@ -16,11 +16,21 @@ from django.views.generic.list import ListView
 from itertools import chain
 from plans.importer import import_name
 from plans.mixins import LoginRequired
-from plans.models import UserPlan, PlanPricing, Plan, Order, BillingInfo
+from plans.base.models import (AbstractUserPlan, AbstractPlanPricing, AbstractPlan,
+                               AbstractOrder, AbstractBillingInfo, AbstractQuota,
+                               AbstractInvoice)
 from plans.forms import CreateOrderForm, BillingInfoForm, FakePaymentsForm
-from plans.models import Quota, Invoice
 from plans.signals import order_started
 from plans.validators import plan_validation
+
+
+UserPlan = AbstractUserPlan.get_concrete_model()
+PlanPricing = AbstractPlanPricing.get_concrete_model()
+Plan = AbstractPlan.get_concrete_model()
+Order = AbstractOrder.get_concrete_model()
+BillingInfo = AbstractBillingInfo.get_concrete_model()
+Quota = AbstractQuota.get_concrete_model()
+Invoice = AbstractInvoice.get_concrete_model()
 
 
 class AccountActivationView(LoginRequired, TemplateView):
@@ -483,4 +493,3 @@ class FakePaymentsView(LoginRequired, SingleObjectMixin, FormView):
             self.object.status = form['status'].value()
             self.object.save()
             return HttpResponseRedirect(reverse('order_payment_failure', kwargs={'pk': self.object.pk}))
-
